requirements.txt
tornado
opencv-python
numpy
Pillow
uvloop